factoring common functionality between Rascal LSP and Parametric DSL with respect to code actions
