+++ rascal-lsp/src/main/java/org/rascalmpl/vscode/lsp/util/CodeActions.java
@@ -27,27 +27,76 @@
 package org.rascalmpl.vscode.lsp.util;
 
 import java.util.Arrays;
+import java.util.List;
+import java.util.Objects;
+import java.util.concurrent.CompletableFuture;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.eclipse.lsp4j.CodeAction;
 import org.eclipse.lsp4j.CodeActionKind;
+import org.eclipse.lsp4j.CodeActionParams;
 import org.eclipse.lsp4j.Command;
+import org.eclipse.lsp4j.Diagnostic;
 import org.eclipse.lsp4j.WorkspaceEdit;
+import org.eclipse.lsp4j.jsonrpc.messages.Either;
 import org.rascalmpl.values.IRascalValueFactory;
 import org.rascalmpl.vscode.lsp.BaseWorkspaceService;
 import org.rascalmpl.vscode.lsp.IBaseTextDocumentService;
 import org.rascalmpl.vscode.lsp.parametric.model.RascalADTs;
 
+import com.google.gson.JsonPrimitive;
+
 import io.usethesource.vallang.IConstructor;
 import io.usethesource.vallang.IList;
 import io.usethesource.vallang.IString;
+import io.usethesource.vallang.IValue;
 import io.usethesource.vallang.IWithKeywordParameters;
 
 /**
  * Reusable utilities for code actions and commands (maps between Rascal and LSP world)
  */
 public class CodeActions {
-    public static CodeAction constructorToCodeAction(IBaseTextDocumentService doc, String dedicatedLanguageName, String languageName, IConstructor codeAction) {
+    /**
+     *  Makes a future stream for filtering out the "fixes" that were optionally sent along with earlier diagnostics
+     *  and which came back with the codeAction's list of relevant (in scope) diagnostics.
+     *
+     * @param params        diagnostics directly from the client (holding embedded action terms)
+     * @param actionParser  provides the parser with a scope that imports the right definitions of Command terms.
+     * @return              a future stream of parsed and type-checked Rascal CodeAction terms.
+     */
+    public static CompletableFuture<Stream<IValue>> extractActionsFromDiagnostics(CodeActionParams params, Function<String, CompletableFuture<IList>> actionParser) {
+        final var emptyListFuture = CompletableFuture.completedFuture(IRascalValueFactory.getInstance().list());
+
+        return params.getContext().getDiagnostics()
+                .stream()
+                .map(Diagnostic::getData)
+                .filter(Objects::nonNull)
+                .filter(JsonPrimitive.class::isInstance)
+                .map(JsonPrimitive.class::cast)
+                .map(JsonPrimitive::getAsString)
+                // this is the "magic" resurrection of command terms from the JSON data field
+                .map(actionParser)
+                // this serializes the stream of futures and accumulates their results as a flat list again
+                .reduce(emptyListFuture, (acc, next) -> acc.thenCombine(next, IList::concat))
+                .thenApply(IList::stream)
+            ;
+    }
+
+    /* merges two streams of CodeAction terms and then converts them to LSP objects */
+    public static CompletableFuture<List<Either<Command, CodeAction>>> mergeAndConvertCodeActions(IBaseTextDocumentService doc, String dedicatedLanguageName, String languageName, CompletableFuture<Stream<IValue>> quickfixes, CompletableFuture<Stream<IValue>> codeActions) {
+         return codeActions.thenCombine(quickfixes, (actions, quicks) ->
+                Stream.concat(quicks, actions)
+                    .map(IConstructor.class::cast)
+                    .map(cons -> constructorToCodeAction(doc, dedicatedLanguageName, languageName, cons))
+                    .map(cmd  -> Either.<Command,CodeAction>forRight(cmd))
+                    .collect(Collectors.toList())
+            );
+    }
+
+    private static CodeAction constructorToCodeAction(IBaseTextDocumentService doc, String dedicatedLanguageName, String languageName, IConstructor codeAction) {
         IWithKeywordParameters<?> kw = codeAction.asWithKeywordParameters();
         IConstructor command = (IConstructor) kw.getParameter(RascalADTs.CodeActionFields.COMMAND);
         IString title = (IString) kw.getParameter(RascalADTs.CodeActionFields.TITLE);
